Updated index.ts
